typings clean up
