+++ src/__tests__/register.test.tsx
@@ -3,9 +3,9 @@ import fetchMock from 'jest-fetch-mock';
 import { signIn } from 'next-auth/react';
 import { randomUUID } from 'crypto';
 import RegisterPage from '@/pages/register';
-import { UpdatedResponse } from '@/features/common';
+import { UpdatedResponse } from '@/features/common/commonTypes';
 import { RegisterFormValues } from '@/features/account/RegisterForm/RegisterForm';
-import { queryWrapper } from './test-utils';
+import { queryWrapper } from '@/utils/test-utils';
 
 const fillRegisterForm = (
     values: Omit<RegisterFormValues, 'confirmPassword'>
