+++ src/features/manage/hooks/useChangePassword.ts
@@ -1,13 +1,8 @@
 import { useMutation } from '@tanstack/react-query';
-import { UpdatedResponse } from '@/features/common';
+import { UpdatedResponse } from '@/features/common/commonTypes';
+import { ChangePasswordRequest } from '../manageTypes';
 import { fetcher } from '@/utils/fetch';
 import { config } from '@/utils/config';
-
-export type ChangePasswordRequest = {
-    currentPassword: string;
-    newPassword: string;
-    version?: number;
-};
 
 const changePassword = (request: ChangePasswordRequest) =>
     fetcher<UpdatedResponse>(`${config.API_URL}/manage/change-password`, {
