+++ src/features/manage/hooks/useDeleteAccount.ts
@@ -1,9 +1,8 @@
-import { UpdatedResponse } from '@/features/common';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { UpdatedResponse } from '@/features/common/commonTypes';
+import { DeleteAccountRequst } from '../manageTypes';
 import { fetcher } from '@/utils/fetch';
 import { config } from '@/utils/config';
-
-export type DeleteAccountRequst = { version?: number };
 
 const deleteAccount = (request: DeleteAccountRequst) =>
     fetcher<UpdatedResponse>(`${config.API_URL}/manage`, {
