+++ src/features/user/hooks/useLockUser.ts
@@ -1,9 +1,8 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { UpdatedResponse } from '@/features/common';
+import { UpdatedResponse } from '@/features/common/commonTypes';
+import { LockUserRequest } from '../userTypes';
 import { fetcher } from '@/utils/fetch';
 import { config } from '@/utils/config';
-
-export type LockUserRequest = { version?: number };
 
 const lockUser = (id: string, request: LockUserRequest) =>
     fetcher<UpdatedResponse>(`${config.API_URL}/user/${id}/lock`, {
